backend/app/requirements.txt
fastapi
uvicorn
openai
python-dotenv
apscheduler
robin-stocks